Properly rename custom_button.dart
