+++ lib/screens/sign_up_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:movie_rating_app/screens/sign_in_screen.dart';
-import 'package:movie_rating_app/widgets/custm_button.dart';
+import 'package:movie_rating_app/widgets/custom_button.dart';
 import 'package:movie_rating_app/widgets/custom_text_field.dart';
 
 class SignUpScreen extends StatefulWidget {
